moved SVSimulator gate application to ext.rs
