+++ quasim/src/ext.rs
@@ -142,6 +142,128 @@ pub fn collapse(state: &[Complex<f64>]) -> usize {
     dist.sample(&mut rng)
 }
 
+/// Checks that all control bits are 1
+fn controls_active(i: usize, controls: QBits) -> bool {
+    let control_mask = controls.get_bitstring();
+    (i & control_mask) == control_mask
+}
+
+/// Checks that all target bits are 0
+fn is_block_base(i: usize, targets: QBits) -> bool {
+    let target_mask = targets.get_bitstring();
+    (i & target_mask) == 0
+}
+
+// 0 1
+// 1 0
+#[inline(always)]
+fn apply_x(state_vector: &mut DVector<Complex<f64>>, base_index: usize, target: QBits) {
+    state_vector
+        .as_mut_slice()
+        .swap(base_index, base_index | target.get_bitstring());
+}
+
+// 0 -i
+// i  0
+#[inline(always)]
+fn apply_y(state_vector: &mut DVector<Complex<f64>>, base_index: usize, target: QBits) {
+    let flipped_index = base_index | target.get_bitstring();
+    let state = state_vector.as_mut_slice();
+    let a = state[base_index];
+    let b = state[flipped_index];
+
+    state[base_index] = cart!(b.im, -b.re);
+    state[flipped_index] = cart!(-a.im, a.re);
+}
+
+// 1  0
+// 0 -1
+#[inline(always)]
+fn apply_z(state_vector: &mut DVector<Complex<f64>>, base_index: usize, target: QBits) {
+    let i = base_index | target.get_bitstring();
+    let amp = &mut state_vector[i];
+
+    amp.re = -amp.re;
+    amp.im = -amp.im;
+}
+
+#[inline(always)]
+fn apply_h(state_vector: &mut DVector<Complex<f64>>, base_index: usize, target: QBits) {
+    let flipped_index = base_index | target.get_bitstring();
+    let state = state_vector.as_mut_slice();
+    let a = state[base_index];
+    let b = state[flipped_index];
+    let inv_sqrt2 = 1.0 / std::f64::consts::SQRT_2;
+
+    state[base_index] = (a + b) * inv_sqrt2;
+    state[flipped_index] = (a - b) * inv_sqrt2;
+}
+
+#[inline(always)]
+fn apply_s(state_vector: &mut DVector<Complex<f64>>, base_index: usize, target: QBits) {
+    let i = base_index | target.get_bitstring();
+    let amp = state_vector[i];
+
+    state_vector[i].re = -amp.im;
+    state_vector[i].im = amp.re;
+}
+
+#[inline(always)]
+fn apply_swap(state_vector: &mut DVector<Complex<f64>>, base_index: usize, targets: QBits) {
+    let t0 = targets.get_indices()[0];
+    let t1 = targets.get_indices()[1];
+
+    let i01 = base_index | (1 << t0);
+    let i10 = base_index | (1 << t1);
+
+    state_vector.as_mut_slice().swap(i01, i10);
+}
+
+#[inline(always)]
+fn apply_unitary2(
+    state_vector: &mut DVector<Complex<f64>>,
+    base_index: usize,
+    u: &Matrix2<Complex<f64>>,
+    target: QBits,
+) {
+    let flipped_index = base_index | target.get_bitstring();
+    let a = state_vector[base_index];
+    let b = state_vector[flipped_index];
+
+    state_vector[base_index] = u[(0, 0)] * a + u[(0, 1)] * b;
+    state_vector[flipped_index] = u[(1, 0)] * a + u[(1, 1)] * b;
+}
+
+pub fn apply_gate(state_vector: &mut DVector<Complex<f64>>, gate: &Gate) {
+    let controls = gate.get_control_bits();
+    let targets = gate.get_target_bits();
+    let n = state_vector.len();
+
+    // No parallelization
+    // State vector is length 2^n , n=num qubits
+    for i in 0..n {
+        if !is_block_base(i, targets) {
+            continue;
+        }
+
+        if !controls_active(i, controls) {
+            continue;
+        }
+
+        match gate.get_type() {
+            GateType::X => apply_x(state_vector, i, targets),
+            GateType::Y => apply_y(state_vector, i, targets),
+            GateType::Z => apply_z(state_vector, i, targets),
+            GateType::H => apply_h(state_vector, i, targets),
+            GateType::S => apply_s(state_vector, i, targets),
+            GateType::SWAP => apply_swap(state_vector, i, targets),
+            GateType::U(theta, phi, lambda) => {
+                apply_unitary2(state_vector, i, &get_u_matrix2(theta, phi, lambda), targets)
+            }
+        }
+    }
+}
+
 /// # measure_state_vector
 /// Returns a measurement and updates the state vector.
 pub fn measure_state_vector(
